requirements.txt
asgiref==3.6.0
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
flake8-html==0.4.3
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
